+++ app/src/androidTest/java/com/swent/assos/AssoDetailsTest.kt
@@ -94,23 +94,6 @@ class AssoDetailsTest : SuperTest() {
   }
 
   @Test
-  fun testNavigationToCommitteeDetails() {
-    run {
-      ComposeScreen.onComposeScreen<AssoDetailsScreen>(composeTestRule) {
-        step("Check if the committee button is displayed") {
-          composeTestRule.waitUntil(
-              condition = { composeTestRule.onNodeWithText("The Committee").isDisplayed() },
-              timeoutMillis = 10000)
-        }
-
-        step("Click on the committee button") {
-          composeTestRule.onNodeWithText("The Committee").performClick()
-        }
-      }
-    }
-  }
-
-  @Test
   fun followAsso() {
     DataCache.currentUser.value.following = emptyList()
     run {
